LayoutTests/http/tests/security/seamless/seamless-cross-origin-expected.txt

FAIL iframe1.seamless should be true (of type boolean). Was undefined (of type undefined).
FAIL iframe2.seamless should be true (of type boolean). Was undefined (of type undefined).
PASS window.getComputedStyle(iframe1).width is "200px"
PASS window.getComputedStyle(iframe1).height is "100px"
PASS window.getComputedStyle(iframe2).width is "300px"
PASS window.getComputedStyle(iframe2).height is "150px"
